Move files + create Menu component
