+++ src/App.jsx
@@ -1,13 +1,17 @@
 import React, { Component } from 'react';
 
-import Scene from './Scene.jsx'
+import Scene from './scene/Scene.jsx';
+import Menu from './menu/Menu.jsx';
 
 class App extends Component {
   render() {
     return (
-      <Scene />
+      <React.Fragment>
+        <Scene />
+        <Menu />
+      </React.Fragment>
     );
+
   }
 }
-
 export default App;
