+++ src/menu/Menu.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Menu() {
+  return <h1>Menu</h1>;
+};
